# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Tracker for Products, Recipes, and Orders",
    "version" : "Layout",
    "title" : "Elevated Inventory Tracker",
    "contact" : {
      "email" : "joshydenberg@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/yjosh/ElevatedV3/1.0.0",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/recipes/{recipeId}" : {
      "get" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives a List of materials for the provided recipeIds",
        "description" : "Search for a specific recipe By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "searchRecipes",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Found recipe(s)",
            "schema" : {
              "type" : "object",
              "properties" : {
                "recipeId" : {
                  "type" : "string"
                },
                "materials" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Material"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Unable to find valid recipe"
          }
        }
      },
      "put" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Edit a specific recipe",
        "description" : "Edit a specific recipe",
        "operationId" : "editRecipe",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "materialIds",
          "description" : "list of materialIds",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "recipe object",
            "schema" : {
              "$ref" : "#/definitions/Recipe"
            }
          },
          "400" : {
            "description" : "Unable to find recipes"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins", "employees", "manager", "technical" ],
        "summary" : "delete a recipe",
        "description" : "Deletes a recipe from the system",
        "operationId" : "deleteRecipe",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "recipe object that was deleted",
            "schema" : {
              "$ref" : "#/definitions/Recipe"
            }
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "unable to delete object"
          }
        }
      }
    },
...
    "/recipes/{recipeId}/material" : {
      "get" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives the materials for a specific recipe",
        "description" : "Retreives the material for a specific recipe By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "searchMaterialByRecipe",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Found Valid list of recipes",
            "schema" : {
              "$ref" : "#/definitions/Material"
            }
          },
          "400" : {
            "description" : "Unable to find valid recipe"
          }
        }
      },
      "put" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives the material for a specific recipe",
        "description" : "Retreives the material for a specific recipe By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "editRecipeByRecipe",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Edited material for recipeId",
            "schema" : {
              "type" : "object",
              "properties" : {
                "recipeId" : {
                  "type" : "string"
                },
                "material" : {
                  "$ref" : "#/definitions/Material"
                }
              }
            }
          },
          "400" : {
            "description" : "Unable to find valid recipe"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins", "technical" ],
        "summary" : "Removes a recipe from a material",
        "description" : "Deletes an item to the system",
        "operationId" : "DeleteRecipe",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "recipeId",
          "in" : "path",
          "description" : "Id of the recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "recipeId",
          "in" : "query",
          "description" : "Id of the recipe to Delete",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Deleted material from recipe",
            "schema" : {
              "type" : "object",
              "properties" : {
                "recipeId" : {
                  "type" : "string"
                },
                "productId" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "invalid input, object does not exist"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Recipe" : {
      "type" : "object",
      "required" : [ "recipeId", "unit", "value" ],
      "properties" : {
        "productId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "recipeId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "materials" : {
          "type" : "object",
          "properties" : { }
        },
        "unit" : {
          "type" : "string",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "value" : {
          "type" : "number"
        }
      }
    },
...
    "Material" : {
      "type" : "object",
      "properties" : {
        "materialId" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "stockQuanitity" : {
          "type" : "number"
        },
        "unit" : {
          "type" : "string"
        },
        "supplierId" : {
          "type" : "string"
        }
      }
    },
...
  }
}